requirements.txt
-i http://mirrors.aliyun.com/pypi/simple/
--trusted-host mirrors.aliyun.com
beautifulsoup4>=4.9.0
bs4>=0.0.1
certifi>=2020.6.20
chardet>=3.0.4
click>=7.0
cssselect>=1.1.0; python_version != '3.3.*'
dill>=0.3.0
easyutils>=0.1.7
flask>=1.1.0
idna>=2.10
itsdangerous>=1.1.0
jinja2>=2.11.0
lxml>=4.6.0
markupsafe>=1.1.0
numpy>=1.19.0; python_version >= '3.6'
pandas>=1.1.0
pillow>=8.0.0
pyperclip>=1.8.0
pyquery>=1.4.0; python_version != '3.0.*'
pytesseract>=0.3.7
python-dateutil>=2.8.0
python-xlib>=0.29; sys_platform != 'win32'
pytz>=2020.1
pywinauto>=0.6.8; sys_platform == 'win32'
requests>=2.25.0
six>=1.15.0
urllib3>=1.26.0; python_version != '3.1.*'
werkzeug>=1.0.0

